Docs and examples updates
