+++ Cargo.toml
@@ -3,6 +3,5 @@
 members = [
     "src/models",
     "src/client",
-    "src/hyper",
-    "src/examples",
+    "src/hyper"
 ]
