+++ README.md
@@ -9,13 +9,19 @@ Slack Morphism is a modern client library for Slack Web/Events API/Sockets Mode 
 Please follow to the official website: https://slack-rust.abdolence.dev
 
 ## Examples
-https://github.com/abdolence/slack-morphism-rust/tree/master/src/examples/src
 
-The example bot requires to work the following environment variables (from your Slack bot profile in api.slack.com):
-- `SLACK_CLIENT_ID`, `SLACK_CLIENT_SECRET`, `SLACK_BOT_SCOPE`, `SLACK_REDIRECT_HOST` - for OAuth
+https://github.com/abdolence/slack-morphism-rust/tree/master/src/hyper/examples
+
+The examples require to work the following environment variables (from your Slack bot profile in api.slack.com):
 - `SLACK_TEST_TOKEN` - for Slack client example
 - `SLACK_TEST_APP_TOKEN` - for Slack client with Socket Mode example
-- `SLACK_SIGNING_SECRET` for all routes
+- `SLACK_CLIENT_ID`, `SLACK_CLIENT_SECRET`, `SLACK_BOT_SCOPE`, `SLACK_REDIRECT_HOST` - for OAuth routes for Events API example
+- `SLACK_SIGNING_SECRET` for all routes for Events API example
+
+To run example use with environment variables:
+```
+# SLACK_... cargo run --example <client|events_api_server|socket_mode>
+```
 
 Routes for this example are available on http://<your-host>:8080:
