--- src/examples/Cargo.toml
@@ -1,25 +1,0 @@
-[package]
-name = "slack-morphism-examples"
-version = "0.10.0"
-authors = ["Abdulla Abdurakhmanov <abdulla@latestbit.com>"]
-edition = "2018"
-publish = false
-
-[dependencies]
-slack-morphism-models = { path = "../models", version = "^0.10.0"}
-slack-morphism-hyper = { path = "../hyper", version = "^0.10.0"}
-slack-morphism = { path = "../client", version = "^0.10.0"}
-serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0"
-serde_with = "1.4"
-futures = "0.3"
-tokio = { version = "1.0", features = ["full"] }
-chrono = { version = "0.4", features = ["serde"] }
-hyper = { version ="0.14", features = ["full"] }
-log = "0.4"
-fern = { version = "0.6", features = ["colored"] }
-rsb_derive = "0.3"
-http = "0.2"
-
-[package.metadata.release]
-disable-tag=true
